fix packets, inventory is now loading correctly

+++ src/ChickenAPI/Entities/Player/Extensions/LevPacketExtension.cs
@@ -18,15 +18,15 @@ namespace ChickenAPI.Game.Entities.Player
             return new LevPacket
             {
                 Level = player.Experience.Level,
-                LevelXp = player.Experience.LevelXp,
+                LevelXp = (int)player.Experience.LevelXp,
                 JobLevel = exp.JobLevel,
-                JobLevelXp = exp.JobLevelXp,
+                JobLevelXp = (int)exp.JobLevelXp,
                 LevelXpMax = Algorithm.GetLevelXp(player.Character.Class, exp.Level),
                 JobLevelXpMax = Algorithm.GetLevelXp(player.Character.Class, exp.JobLevel),
                 HeroLevel = exp.HeroLevel,
                 Reputation = player.Character.Reput,
                 Cp = player.GetCp(),
-                HeroLevelXp = exp.HeroLevelXp,
+                HeroLevelXp = (int)exp.HeroLevelXp,
                 HeroLevelXpMax = Algorithm.GetHeroLevelXp(player.Character.Class, exp.HeroLevel),
             };
         }
